ccd_script
